refactor: move isErrRetryable function to utils
