+++ packages/analytics-js/src/components/configManager/ConfigManager.ts
@@ -140,6 +140,7 @@ class ConfigManager implements IConfigManager {
    */
   processConfig(response?: SourceConfigResponse | string, details?: ResponseDetails) {
     // TODO: add retry logic with backoff based on rejectionDetails.xhr.status
+    // We can use isErrRetryable utility method
     if (!response) {
       this.onError(SOURCE_CONFIG_FETCH_ERROR(details?.error));
       return;
